+++ components/camera/camera.go
@@ -44,14 +44,6 @@ func init() {
 const (
 	// SubtypeName is a constant that identifies the camera resource subtype string.
 	SubtypeName = "camera"
-
-	// TimeRequestedMetadataKey is optional metadata in the gRPC response header that correlates
-	// to the time right before the point cloud was captured.
-	TimeRequestedMetadataKey = "viam-time-requested"
-
-	// TimeReceivedMetadataKey is optional metadata in the gRPC response header that correlates
-	// to the time right after the point cloud was captured.
-	TimeReceivedMetadataKey = "viam-time-received"
 )
 
 // API is a variable that identifies the camera resource API.
